+++ src/lib.rs
@@ -2,10 +2,10 @@
 //! and access to various system registers.
 
 #![cfg_attr(not(test), no_std)]
-#![cfg_attr(feature = "const_fn", feature(const_panic))]
-#![cfg_attr(feature = "const_fn", feature(const_mut_refs))]
-#![cfg_attr(feature = "const_fn", feature(const_fn_fn_ptr_basics))]
-#![cfg_attr(feature = "const_fn", feature(const_fn_trait_bound))]
+#![cfg_attr(feature = "const_fn", feature(const_panic))] // Better panic messages
+#![cfg_attr(feature = "const_fn", feature(const_mut_refs))] // GDT add_entry()
+#![cfg_attr(feature = "const_fn", feature(const_fn_fn_ptr_basics))] // IDT new()
+#![cfg_attr(feature = "const_fn", feature(const_fn_trait_bound))] // PageSize marker trait
 #![cfg_attr(feature = "inline_asm", feature(asm))]
 #![cfg_attr(feature = "abi_x86_interrupt", feature(abi_x86_interrupt))]
 #![cfg_attr(docsrs, feature(doc_cfg))]