+++ src/structures/gdt.rs
@@ -112,16 +112,18 @@ impl GlobalDescriptorTable {
     /// * The user must make sure that the entries are well formed
     /// * The provided slice **must not be larger than 8 items** (only up to the first 8 will be observed.)
     #[inline]
-    #[cfg(feature = "const_fn")]
     pub const unsafe fn from_raw_slice(slice: &[u64]) -> GlobalDescriptorTable {
-        assert!(
-            slice.len() <= 8,
-            "initializing a GDT from a slice requires it to be **at most** 8 elements."
-        );
         let next_free = slice.len();
-
         let mut table = [0; 8];
         let mut idx = 0;
+
+        #[cfg(feature = "const_fn")]
+        assert!(
+            next_free <= 8,
+            "initializing a GDT from a slice requires it to be **at most** 8 elements."
+        );
+        #[cfg(not(feature = "const_fn"))]
+        table[next_free]; // Will fail if slice.len() > 8
 
         while idx != next_free {
             table[idx] = slice[idx];